make simulations without failures

+++ org.palladiosimulator.blockchainsystems.core/src/main/kotlin/org/palladiosimulator/blockchainsystems/core/network/LinkThroughput.kt
@@ -11,6 +11,6 @@ import org.palladiosimulator.blockchainsystems.core.common.abstractions.Temporal
  * @author Davis Riedel
  */
 data class LinkThroughput(
-  override val value: Int,
+  override val value: Long,
   override val duration: Long
-) : TemporalValue<Int>
+) : TemporalValue<Long>
